pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-system"
version = "0.1.0"
description = "A production-grade RAG system with hybrid search, reranking, and evaluation"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pydantic-settings>=2.7",
    "python-dotenv>=1.1",
    "langchain-community>=0.3",
    "langchain-text-splitters>=0.3",
    "langchain-huggingface>=0.1",
    "langchain-qdrant>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "pypdf>=5.0",
    "sentence-transformers>=3.3",
    "qdrant-client>=1.13",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.18",

]

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
    "pytest>=8.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
